Refractor all files to ts and fix compilation issues
