+++ nodemon.json
@@ -1,0 +1,12 @@
+{
+  "restartable": "rs",
+  "ignore": [
+    ".git",
+    "node_modules/**/node_modules",
+    "lib/**/*"
+  ],
+  "execMap": {
+    "ts": "node --loader ts-node/esm"
+  },
+  "ext": "ts,js,json"
+}
