+++ package.json
@@ -2,13 +2,10 @@
   "name": "lowry-bot",
   "version": "1.0.0",
   "description": "",
-  "main": "app.js",
+  "main": "./src/app.ts",
   "type": "module",
   "scripts": {
-    "start": "nodemon .",
-    "pm2 start": "pm2 start app.js --watch",
-    "pm2 stop": "pm2 stop app.js",
-    "pm2 log": "pm2 log app.js --lines 100",
+    "start": "nodemon src/app.ts",
     "deploy": "node helperScripts/deploy-commands.js",
     "retract": "node helperScripts/retract-commands.js"
   },
@@ -24,14 +21,21 @@
   "homepage": "https://github.com/Burtorustum/Lowry-Bot#readme",
   "dependencies": {
     "@discordjs/builders": "^1.0.0",
+    "@discordjs/collection": "^1.0.0",
     "@discordjs/rest": "^1.0.0",
     "discord-api-types": "^0.36.3",
     "discord.js": "^14.0.3",
     "dotenv": "^16.0.1",
+    "pino": "^8.3.0",
     "sequelize": "^6.21.3",
-    "sqlite3": "^5.0.9"
+    "sqlite3": "^5.0.10"
   },
   "devDependencies": {
-    "nodemon": "^2.0.19"
+    "@swc/core": "^1.2.218",
+    "@types/node": "^18.0.6",
+    "nodemon": "^2.0.19",
+    "pino-pretty": "^8.1.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.7.4"
   }
 }
