+++ src/database/define-sequelize-model.ts
@@ -1,0 +1,8 @@
+import {Sequelize} from 'sequelize';
+
+export default new Sequelize('database', 'username', 'password', {
+  host: 'localhost',
+  dialect: 'sqlite',
+  logging: false,
+  storage: 'database.sqlite',
+});
